Added PUT and GET api calls to work with stored procedures
When putting, the API will get the user from the database, and use its ID to update a table.
Get results in getting the entire user object from the database.

Need to implement a check in the POST method to make sure a user with inserted mail doesn't exist already.
Also need to add an auto-incrementer for the user_id in the database.